requirements/dev.txt
-r base.txt
django-extensions
django-debug-toolbar
ipdb
